feat: improve GitHub output in case of errors
Groups in the GitHub output are always closed. When an error is found
and printed within a group, it is, therefore, automatically hidden in
the output. This change adds an additional error output after a group
with errors, to makes it easy to find the affected groups.

This commit also cleans up the project setup:
- use uv
- drop poetry
- clean up pyproject.toml to adhere to specs and remove unused things
- use hatch-vcs to parse the release version from the latest Git tag