fix: only use "-dev" version when determining version from latest tag
When the version is determined from a release commit message or a
release branch name, that is the version we want to see in the rule
files. The only case where we need to bump the minor and append "-dev"
is when we determine the version based on the latest tag.